test: Move validating_consumer to test/lib/mutation_assertions.hh
